Allow creating folders recursively
When downloading a drive folder with intermediary directories that
do not contain any files like
root/
└── dir1/
    └── dir2/
    	└── file.ext
the corresponding `drive clone` fails.
Before this commit, `os.mkdir()` was used which fails because it
assumes that `dir1` exists and only creates `dir2`.
After this commit, `os.makedirs()` is used which creates all
required intermediary directories.